[arcane,cuda] Simplifie l'intégration de CUPTI.
Corrige aussi une erreur de compilation si CUPTI n'est pas disponible.
